[Cl_Task_1] move class loader files to separate package
